requirements.txt
fastapi[standard]
python-multipart
Pillow
